requirements.txt
beautifulsoup4==4.6.3
certifi==2018.4.16
chardet==3.0.4
cmake==3.12.0
colorama==0.3.9
defusedxml==0.5.0
Django==2.1
django-allauth==0.36.0
django-rest-auth==0.9.3
django-tastypie==0.14.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gcloud==0.17.0
get==1.0.3
googleapis-common-protos==1.5.3
httpie==0.9.9
httplib2==0.11.3
idna==2.7
jws==0.1.3
lxml==4.2.4
numpy==1.15.1
oauth2client==3.0.0
oauthlib==2.1.0
opencv-python==3.4.2.17
Pillow==5.2.0
pipenv==2018.7.1
post==1.0.2
protobuf==3.6.1
public==1.0.3
pyasn1==0.4.4
pyasn1-modules==0.2.2
Pygments==2.2.0
PyJWT==1.6.4
python-dateutil==2.7.3
python-jwt==2.0.1
python-mimeparse==1.6.0
python3-openid==3.1.0
pytz==2018.5
query-string==1.0.2
request==1.0.2
requests==2.11.1
requests-oauthlib==1.0.0
rsa==3.4.2
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
